feat(item): create integration of delete  item in db by id and avocadoPay gateway

+++ src/main/java/com/italo/catalogy/service/ItemService.java
@@ -117,7 +117,13 @@ public class ItemService {
             this.itemRepository.save(itemModel);
             return;
         }
+        ResponseEntity<ItemAvocadoPayResponseDTO> itemAvocadoPayResponseDTOResponseEntity = this.avocadoPayConfig.deleteItemById(itemModel.getGatewayId());
+
+        if (itemAvocadoPayResponseDTOResponseEntity.getStatusCode()!=HttpStatus.OK)
+            throw new RuntimeException("Deu ruin");
+
         this.itemRepository.deleteById(id);
+
     }
